Fix: Function signature for node indices

+++ KraftonEngine/Source/Engine/Render/Shadow/ShadowAtlasQuadTree.h
@@ -48,10 +48,10 @@ public:
 
 private:
 	// Allocates the node at NodeIdx and returns the corresponding atlas region. Returns invalid region if the node is occupied or too small.
-	FAtlasRegion AllocateNode(uint32 NodeIdx, uint32 RequestedSize);
+	FAtlasRegion AllocateNode(int32 NodeIdx, uint32 RequestedSize);
 
 	// Greedily splits the quadtree to find the best fit for the new shadow map
-	void  Split(float Idx);
+	void  Split(int32 Idx);
 
 	// Ranks the importance of the input light source based on its properties.
 	float EvaluateResolution(const FLightInfo& InLightInfo) const;
